refactor(v4): rename the service modules to lowercase
`src/services/` kept PascalCase filenames from when these modules exported
classes. They export functions now, so they follow the convention the rest
of `src/` uses — `context.ts`, `registry.ts`, `scheduler.ts`,
`mount-strategies.ts`. `Base.ts` stays PascalCase because it exports a
class.

Pure rename, no behaviour change: `git mv` for every module and its spec,
plus the matching import specifiers.
