+++ packages/v4/src/services/service.bench.ts
@@ -6,8 +6,8 @@
  * design questions rather than guard against regressions.
  */
 import { bench, describe } from 'vitest';
-import { BREAKPOINTS } from './ResizeService.js';
-import { createService } from './Service.js';
+import { BREAKPOINTS } from './resize.js';
+import { createService } from './service.js';
 
 /** Keeps each result observable so no benchmark is optimised away. */
 declare global {
